chore(progress-circle): refactor with base class for swan
