+++ packages/progress-circle/src/ProgressCircle.ts
@@ -13,49 +13,19 @@
 import {
     CSSResultArray,
     html,
-    PropertyValues,
-    SizedMixin,
-    SpectrumElement,
     TemplateResult,
 } from '@spectrum-web-components/base';
-import {
-    property,
-    query,
-} from '@spectrum-web-components/base/src/decorators.js';
-import { getLabelFromSlot } from '@spectrum-web-components/shared/src/get-label-from-slot.js';
 import { ifDefined } from '@spectrum-web-components/base/src/directives.js';
+import { ProgressCircleBase } from './ProgressCircleBase.js';
 
 import progressCircleStyles from './progress-circle.css.js';
 
 /**
  * @element sp-progress-circle
  */
-export class ProgressCircle extends SizedMixin(SpectrumElement, {
-    validSizes: ['s', 'm', 'l'],
-}) {
+export class ProgressCircle extends ProgressCircleBase {
     public static override get styles(): CSSResultArray {
         return [progressCircleStyles];
-    }
-
-    @property({ type: Boolean, reflect: true })
-    public indeterminate = false;
-
-    @property({ type: String })
-    public label = '';
-
-    @property({ reflect: true, attribute: 'static-color' })
-    public staticColor?: 'white';
-
-    @property({ type: Number })
-    public progress = 0;
-
-    @query('slot')
-    private slotEl!: HTMLSlotElement;
-
-    private makeRotation(rotation: number): string | undefined {
-        return this.indeterminate
-            ? undefined
-            : `transform: rotate(${rotation}deg);`;
     }
 
     protected override render(): TemplateResult {
@@ -84,61 +54,5 @@ export class ProgressCircle extends SizedMixin(SpectrumElement, {
                 )}
             </div>
         `;
-    }
-
-    protected handleSlotchange(): void {
-        const labelFromSlot = getLabelFromSlot(this.label, this.slotEl);
-        if (labelFromSlot) {
-            this.label = labelFromSlot;
-        }
-    }
-
-    protected override firstUpdated(changes: PropertyValues): void {
-        super.firstUpdated(changes);
-        if (!this.hasAttribute('role')) {
-            this.setAttribute('role', 'progressbar');
-        }
-    }
-
-    protected override updated(changes: PropertyValues): void {
-        super.updated(changes);
-        if (!this.indeterminate && changes.has('progress')) {
-            this.setAttribute('aria-valuenow', '' + this.progress);
-        } else if (this.hasAttribute('aria-valuenow')) {
-            this.removeAttribute('aria-valuenow');
-        }
-        if (changes.has('label')) {
-            if (this.label.length) {
-                this.setAttribute('aria-label', this.label);
-            } else if (
-                changes.get('label') === this.getAttribute('aria-label')
-            ) {
-                this.removeAttribute('aria-label');
-            }
-        }
-
-        if (window.__swc.DEBUG) {
-            if (
-                !this.label &&
-                !this.getAttribute('aria-label') &&
-                !this.getAttribute('aria-labelledby') &&
-                !this.slotEl.assignedNodes().length
-            ) {
-                window.__swc.warn(
-                    this,
-                    '<sp-progress-circle> elements need one of the following to be accessible:',
-                    'https://opensource.adobe.com/spectrum-web-components/components/progress-circle/#accessibility',
-                    {
-                        type: 'accessibility',
-                        issues: [
-                            'value supplied to the "label" attribute, which will be displayed visually as part of the element, or',
-                            'text content supplied directly to the <sp-progress-circle> element, or',
-                            'value supplied to the "aria-label" attribute, which will only be provided to screen readers, or',
-                            'an element ID reference supplied to the "aria-labelledby" attribute, which will be provided by screen readers and will need to be managed manually by the parent application.',
-                        ],
-                    }
-                );
-            }
-        }
     }
 }
